refactor(approval): modularize approval facade
Motivation: The previous approval module interleaved presets, registry

helpers, and redaction utilities which made policy changes risky and

hard to reason about.

Summary: Split the monolith into a small facade package that

re-exports focused modules, removed the legacy file, and updated

affected integration and unit tests to import the new surface.

Implementation: Added dedicated modules for interrupts, presets,

registry policies, and redaction, plus a facade that lazily

imports Inspect internals; refreshed tests to import the package

after installing stubs.

Impact: Approval behavior stays the same while future edits

become easier thanks to clearer structure.

- Ran CI=1 NO_NETWORK=1 PYTHONPATH=src:external/inspect_ai uv run

  --no-project pytest -q tests/unit/inspect_agents/approvals

- Ran CI=1 NO_NETWORK=1 PYTHONPATH=src:external/inspect_ai uv run

  --no-project pytest -q tests/integration/inspect_agents/test_approval_chains.py

- Ran CI=1 NO_NETWORK=1 PYTHONPATH=src:external/inspect_ai uv run

  --no-project pytest -q tests/integration/inspect_agents/test_handoff_exclusivity_integration.py
